+++ onsiteprint-plugin/blocks/event/block-template-parts/list.php
@@ -1,7 +1,8 @@
 <?php
 /* ------------------------------------------------------------------------
  *  Block Part Name: List
- ?  Updated: 2023-04-13 - 19:48 (Y:m:d - H:i)
+ ?  Updated: 2025-06-26 - 11:43 (Y:m:d - H:i)
+ ?  Info: Added Participant Skeleton.
 ---------------------------------------------------------------------------
  #  The Block Part Content
 --------------------------------------------------------------------------- */
@@ -162,9 +163,24 @@
 
     <?php } else { ?>
 
-        <p class="op-flex-col">
-            <span class="op-text">Loading...</span>
-        </p>
+        <article class="op-participant_skeleton op-active op-fade-in">
+            <p class="op-flex-col">
+                <span class="op-shimmer"></span>
+                <span class="op-text">Loading...</span>
+            </p>
+        </article>
+        <article class="op-participant_skeleton op-active op-fade-in">
+            <p class="op-flex-col">
+                <span class="op-shimmer"></span>
+                <span class="op-text">Loading...</span>
+            </p>
+        </article>
+        <article class="op-participant_skeleton op-active op-fade-in">
+            <p class="op-flex-col">
+                <span class="op-shimmer"></span>
+                <span class="op-text">Loading...</span>
+            </p>
+        </article>
 
     <?php